% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ts.R
\name{apply_laplacePoisson}
\alias{apply_laplacePoisson}
\title{Apply the algorithms to make decisions for Thompson sampling Poisson (TS-Poisson) algorithms}
\usage{
apply_laplacePoisson(context, beta_laplacePoisson)
}
\arguments{
\item{context}{context at the current decision time}

\item{beta_laplacePoisson}{the randomly sampled Bayesian estimate}
}
\value{
Intervention option
}
\description{
Apply the algorithms to make decisions for Thompson sampling Poisson (TS-Poisson) algorithms
}
\examples{
apply_laplacePoisson(matrix(1:10, nrow = 2),matrix(11:20, nrow = 5))
}
