% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ts.R
\name{apply_linearTS}
\alias{apply_linearTS}
\title{Apply the algorithms to make decisions for Linear Thompson sampling (TS) algorithms}
\usage{
apply_linearTS(context, beta_linearTS)
}
\arguments{
\item{context}{context at the current decision time}

\item{beta_linearTS}{the randomly sampled Bayesian estimate}
}
\value{
Intervention option
}
\description{
Apply the algorithms to make decisions for Linear Thompson sampling (TS) algorithms
}
\examples{
apply_linearTS(matrix(1:10, nrow = 2),matrix(11:20, nrow = 5))
}
