% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ts.R
\name{apply_normalNB}
\alias{apply_normalNB}
\title{Apply the algorithms to make decisions for Thompson sampling Negative Binomial (TS-NB) algorithms}
\usage{
apply_normalNB(context, beta_normalNB)
}
\arguments{
\item{context}{context at the current decision time}

\item{beta_normalNB}{the randomly sampled Bayesian estimate}
}
\value{
Intervention option
}
\description{
Apply the algorithms to make decisions for Thompson sampling Negative Binomial (TS-NB) algorithms
}
\examples{
apply_normalNB(matrix(1:10, nrow = 2),matrix(11:20, nrow = 5))

}
