% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ts.R
\name{apply_ZINB}
\alias{apply_ZINB}
\title{Apply the algorithms to make decisions for Thompson sampling Zero-inflated Negative Binomial (TS-ZINB) algorithm}
\usage{
apply_ZINB(context, beta_ZINB, gamma_ZINB)
}
\arguments{
\item{context}{context at the current decision time}

\item{beta_ZINB}{the randomly sampled Bayesian estimate for the Poisson component}

\item{gamma_ZINB}{the randomly sampled Bayesian estimate for the zero component}
}
\value{
Intervention option
}
\description{
Apply the algorithms to make decisions for Thompson sampling Zero-inflated Negative Binomial (TS-ZINB) algorithm
}
\examples{
apply_ZINB(matrix(1:10, nrow = 2),matrix(11:20, nrow = 5),matrix(21:30, nrow = 5))
}
