% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ts.R
\name{apply_ZIP}
\alias{apply_ZIP}
\title{Apply the algorithms to make decisions for Thompson sampling Zero-inflated Poisson (TS-ZIP) algorithm}
\usage{
apply_ZIP(context, beta_ZIP, gamma_ZIP)
}
\arguments{
\item{context}{context at the current decision time}

\item{beta_ZIP}{the randomly sampled Bayesian estimate for the Poisson component}

\item{gamma_ZIP}{the randomly sampled Bayesian estimate for the zero component}
}
\value{
Intervention option
}
\description{
Apply the algorithms to make decisions for Thompson sampling Zero-inflated Poisson (TS-ZIP) algorithm
}
\examples{
apply_ZIP(matrix(1:10, nrow = 2),matrix(11:20, nrow = 5),matrix(21:30, nrow = 5))
}
